refactor(web): decompose build detail workspace
