+++ apps/web/src/components/build-details/build-detail-route.tsx
@@ -1,0 +1,8 @@
+import { useParams } from '@tanstack/react-router'
+
+import { BuildDetailPage } from './build-detail-page'
+
+export function BuildDetailRoute() {
+  const { buildId } = useParams({ from: '/builds/$buildId' })
+  return <BuildDetailPage key={buildId} buildId={buildId} />
+}
